Fix last update.
I truly believe that last update should not be localised, and that it
should remain in english.
When the non-serbian speaking user accesses the site, he will have an
easier time to understand when the site was last updated.
Another point is that it messed up the layout less than the localised
version.